+++ client/use-collection.js
@@ -3,22 +3,11 @@ import uniq from 'lodash/uniq.js';
 
 import {computed, reactive} from 'vue';
 import {useRoute} from 'vitepress';
-import encodeTag from './encode-tag.js';
 import {data as collections} from './collections.data.js';
 
 export default function useCollection(type = undefined) {
   const route = useRoute();
-  const urlSearchParams = new URLSearchParams(window.location.search);
-
   const path = route.path;
-  const params = Object.fromEntries(urlSearchParams.entries());
-
-  function isTagSelected(q) {
-    const {tag = '', tags = ''} = params;
-    const list = [...tag.split(','), ...tags.split(',')];
-
-    return list.includes(q) || list.includes(encodeTag(q));
-  }
 
   function findCurrentIndex() {
     const result = pages.findIndex(p => p.url === route.path);
@@ -59,7 +48,9 @@ export default function useCollection(type = undefined) {
   }
 
   // and a selected tag reactive for filtering and that sort of thing
-  const selectedTags = reactive(Object.fromEntries(tags.value.map(tag => ([tag, isTagSelected(tag)]))));
+  // @TODO: it would be great to set selectedTags from query params instead of just false but when we try to do it the
+  // obvious way we get a hydration mismatch which breaks filtering on prod
+  const selectedTags = reactive(Object.fromEntries(tags.value.map(tag => ([tag, false]))));
   const selectedTagsList = computed(() => Object.entries(selectedTags)
     .filter(pair => pair[1] === true)
     .map(pair => pair[0]));